pyproject.toml
[project]
name = "multimodal-economic-rag-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.8.3",
    "fastapi==0.115.12",
    "uvicorn[standard]==0.34.2",
    "httpx>=0.28.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.17.1",
    "openai==1.82.0",
    "pydantic==2.11.3",
    "pydantic-settings==2.9.1",
    "langsmith>=0.7.22",
    "langgraph==1.1.3",
    "loguru>=0.7.0",
    "anyio>=4.0",
    "autoawq>=0.2.9",
    "autoawq-kernels>=0.0.9",
    "transformers>=5.5.4",
    "huggingface-hub>=1.7.2",
    "accelerate>=1.13.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = '--import-mode=importlib -m "not integration"'
markers = [
    "integration: requires Docker, pinned service images, or GPU-backed runtime and is skipped by default",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.35.0",
]
